carousel_tool/requirements.txt
anthropic>=0.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
trafilatura>=1.8.0
python-docx>=1.1.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-generativeai>=0.3.0
schedule>=1.2.0
python-dotenv>=1.0.0
flask>=3.0.0
Pillow>=10.0.0
